+++ src/workload-orchestration/azext_workload_orchestration/aaz/latest/workload_orchestration/configuration/_config_set.py
@@ -29,9 +29,9 @@ class ShowConfig2(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-08-01-preview",
+        "version": "2025-08-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/Microsoft.Edge/solutions/{}", "2024-08-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/Microsoft.Edge/solutions/{}", "2025-08-01"],
         ]
     }
 
@@ -226,7 +226,7 @@ class ShowConfig2(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-06-01-preview",
+                    "api-version", "2025-08-01",
                     required=True,
                 ),
             }