+++ src/workload-orchestration/azext_workload_orchestration/aaz/latest/workload_orchestration/solution_template/_remove_version.py
@@ -21,9 +21,9 @@ class RemoveVersion(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-06-01",
+        "version": "2025-08-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.edge/solutiontemplates/{}/removeversion", "2025-06-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/Microsoft.Edge/solutiontemplates/{}/removeversion", "2025-08-01"],
         ]
     }
 
@@ -139,7 +139,7 @@ class RemoveVersion(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-06-01",
+                    "api-version", "2025-08-01",
                     required=True,
                 ),
             }